remove separate structs for sentinel in directory service
